+++ cliquet/initialization.py
@@ -21,7 +21,6 @@ from cliquet import errors
 from cliquet import logger
 from cliquet import utils
 from cliquet import statsd
-from cliquet import authorization
 
 from pyramid.events import NewRequest, NewResponse
 from pyramid.httpexceptions import HTTPTemporaryRedirect, HTTPGone
@@ -80,8 +79,6 @@ def setup_authentication(config):
     from configuration.
     """
     config.include('pyramid_multiauth')
-    # By default, permissions are handled dynamically.
-    config.set_default_permission(authorization.DYNAMIC)
 
     # Track policy used, for prefixing user_id and for logging.
     def on_policy_selected(event):