+++ cliquet/tests/testapp/views.py
@@ -1,5 +1,4 @@
 from cliquet import resource
-from cliquet import authorization
 import colander
 
 
@@ -7,11 +6,11 @@ class MushroomSchema(resource.ResourceSchema):
     name = colander.SchemaNode(colander.String())
 
 
-@resource.register(factory=authorization.RouteFactory)
+@resource.register()
 class Mushroom(resource.BaseResource):
     mapping = MushroomSchema()
 
 
-@resource.register(factory=authorization.RouteFactory)
+@resource.register()
 class Toadstool(resource.ProtectedResource):
     mapping = MushroomSchema()