+++ CHANGELOG.rst
@@ -11,7 +11,16 @@ This document describes changes between each past release.
 - Provide details about existing record in ``412`` error responses
   (fixes mozilla-services/kinto#122)
 
-2.2.1 (unreleased)
+**Bug fixes**
+
+- Default resources do accept ``permissions`` attribute in payload anymore
+- Default resources do not require a root factory (fixes #348)
+- Default resources do not hit the permission backend anymore
+- Default viewset was split and does not handle permissions anymore (fixes #322)
+- Permissions on views is now set only on resources
+
+
+2.2.1 (2015-07-06)
 ------------------
 
 **Bug fixes**